+++ lib/generators.dart
@@ -1,4 +1,5 @@
 import 'dart:convert';
+import 'dart:html';
 
 import 'package:dart_style/dart_style.dart';